Add: Hostname and port usable/accessible in Database.Status fields.
Realized consuming apps won't be able to get hostname and port coordinates where adminconnection is separated across a namespace boundary and the deploy/user is not allowed via RBAC. This will help and isn't the sensitive parts.

Partial work on #28. Need to check if asking database for RANDOM PASSWORD will allow adherence to business-specific rules (or if I want to allow operator config to drive it).