Fixes #31: Minor discrepencies between MySQL and MariaDB
Putting the onus on the user to generate right combination
of ClearText and changes/setting AuthPlugin.

Also, stop reflecting back GRANT USAGE entry from
SHOW GRANTS. Unclear if hashed password returned by
MariaDB in this case is concerning or not, but no value in
it.

Password was always being monitored and updated in the
database. Removed TODO.